Prototype of 23kbps encoding
This commit adds prototype waveform generating code (in Python) and a
corresponding receiver on the Apple IIe side. It is intended as a
base for experimentation, as the waveform was calibrated to work well
for one machine, and more testing is needed to determine if it needs
recalibration on a machine-by-machine basis.

See #4 for more discussion.